@@ -12,19 +12,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^0.803.25",
-    "@angular/animations": "^8.2.4",
-    "@angular/cdk": "^8.1.4",
-    "@angular/cli": "^8.3.25",
-    "@angular/common": "^8.2.4",
-    "@angular/compiler": "^8.2.4",
-    "@angular/compiler-cli": "^8.2.4",
-    "@angular/core": "^8.2.4",
-    "@angular/forms": "^8.2.4",
-    "@angular/material": "^8.1.4",
-    "@angular/platform-browser": "^8.2.4",
-    "@angular/platform-browser-dynamic": "^8.2.4",
-    "@angular/router": "^8.2.4",
+    "@angular-devkit/build-angular": "~0.900.2",
+    "@angular/animations": "^9.0.1",
+    "@angular/cdk": "^9.0.0",
+    "@angular/cli": "^9.0.2",
+    "@angular/common": "^9.0.1",
+    "@angular/compiler": "^9.0.1",
+    "@angular/compiler-cli": "^9.0.1",
+    "@angular/core": "^9.0.1",
+    "@angular/forms": "^9.0.1",
+    "@angular/localize": "^9.0.1",
+    "@angular/material": "^9.0.0",
+    "@angular/platform-browser": "^9.0.1",
+    "@angular/platform-browser-dynamic": "^9.0.1",
+    "@angular/router": "^9.0.1",
     "@auth0/angular-jwt": "^4.0.0",
     "@fortawesome/angular-fontawesome": "^0.5.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.27",
@@ -47,21 +48,21 @@
     "filepond-plugin-image-preview": "^4.6.1",
     "filepond-plugin-image-resize": "^2.0.4",
     "filepond-plugin-image-transform": "^3.7.1",
-    "hammerjs": "^2.0.8",
     "lodash": "^4.17.15",
     "ngx-filepond": "^5.0.1",
     "rxjs": "~6.5.2",
     "socket.io": "^2.3.0",
     "sweetalert2": "^9.7.2",
-    "typescript": "~3.4.5",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "typescript": "~3.7.5",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular/language-service": "^8.2.4",
+    "@angular/language-service": "^9.0.1",
     "@ngrx/schematics": "^8.3.0",
     "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.0.10",
+    "@types/node": "^12.11.1",
     "@types/socket.io": "^2.1.4",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -73,7 +74,6 @@
     "prettier": "^1.19.1",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslib": "^1.10.0",
     "tslint": "~5.11.0",
     "tslint-config-prettier": "^1.18.0"
   }
