@@ -40,7 +40,7 @@ import {
   exports: [LoginComponent, RegisterUserComponent]
 })
 export class AuthModule {
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<AuthModule> {
     return {
       ngModule: AuthModule,
       providers: [AuthenticationService]
