@@ -1,5 +1,6 @@
 import { Component, OnInit, OnDestroy, ViewChild } from '@angular/core';
-import { MatDialog, MatTableDataSource, MatTable } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
+import { MatTableDataSource, MatTable } from '@angular/material/table';
 import { ActivatedRoute } from '@angular/router';
 import Swal from 'sweetalert2';
 import { Role, Privilege, RolePrivilege } from '../../_models';
@@ -28,7 +29,7 @@ export class RoleComponent implements OnInit, OnDestroy {
   editRolePrivilege$: Observable<boolean>;
   displayedColumns = ['title', 'view', 'add', 'edit', 'delete'];
 
-  @ViewChild(MatTable, { static: false }) privilegesTable: MatTable<any>;
+  @ViewChild(MatTable) privilegesTable: MatTable<any>;
 
   private unsubscribe: Subject<void> = new Subject();
 
